+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## v2.0.0 ()
 - The collection of units of measure that a given quantity supports is now a static property of physical quantity classes, as opposed to a property of each individual value object.  This means that adding units of measure to Length, say, will make those units available to all Length instances.  This has the effect of greatly reducing UnitOfMeasure object count in situations where large numbers of PhysicalQuantity objects are instantiated.
   - As part of this change, default units of measure are no longer added in the quantity class constructor, and are instead handled in a initializeUnitsOfMeasure() static method.
-  - registerUnitOfMeasure() and getSupportedUnits() are now static methods.
+  - registerUnitOfMeasure() and getSupportedUnitNames() are now static methods.
 - Physical quantities now have an abstract interface to implement - PhysicalQuantityInterface
 - Make sure the native unit of measure idea is still necessary
 
@@ -12,7 +12,7 @@
 - Add support for automatically generating metric-prefixed units from a single base unit
 - Throw an exception on registering units of measure with names or aliases that collide with existing units on that quantity
 - Disallow values that aren't numerical and unit names that aren't strings (ie, type checking)
-- Add new method getSupportedUnits() to Physical Quantities
+- Add new method getSupportedUnitNames() to Physical Quantities
 - Add new method getAliases() to Units of Measure
 - Add arcminute and arcsecond as angle units
 - Add all SI prefix units to meters, kilograms, seconds, amps, kelvin, candela, pascal, bar, radians, degrees, and arcseconds